Add alert for delete page button
When user clicks on the delete page button, an alert pops up indicating
that the delete button must be pressed again in order to proceed with
deleting the page. The delete button then turns red to indicate that
clicking it will immediately delete the page. If user clicks on any
other button or changes the text in the textarea (while the delete
button is red), the delete will be cancelled and user will be alerted
again when clicking the delete button the next time.